Inline launcher submod
