+++ oak_functions_launcher/benches/integration_benches.rs
@@ -62,7 +62,7 @@ fn run_bench(b: &mut Bencher, config: &OakFunctionsTestConfig) {
         oak_functions_test_utils::build_rust_crate_enclave("oak_functions_enclave_app")
             .expect("Failed to build oak_functions_enclave_app");
 
-    let params = launcher::virtualized::Params {
+    let params = launcher::Params {
         enclave_binary: workspace_path(&[
             "oak_restricted_kernel_bin",
             "target",
